Move common R/G functionality to base class
