+++ include/Basis/SpinorBasis/SimpleSpinorBasis.hpp
@@ -1,0 +1,115 @@
+// This file is part of GQCG-gqcp.
+// 
+// Copyright (C) 2017-2019  the GQCG developers
+// 
+// GQCG-gqcp is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Lesser General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+// 
+// GQCG-gqcp is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Lesser General Public License for more details.
+// 
+// You should have received a copy of the GNU Lesser General Public License
+// along with GQCG-gqcp.  If not, see <http://www.gnu.org/licenses/>.
+// 
+#pragma once
+
+
+#include "Basis/TransformationMatrix.hpp"
+
+
+namespace GQCP {
+
+
+/**
+ *  A class that represents a spinor basis that has no internal structure (hence 'simple'). It is used as a base class for restricted and generalized spinor bases because they admit some common functionality
+ * 
+ *  @tparam _ExpansionScalar        the scalar type of the expansion coefficients
+ */
+template <typename _ExpansionScalar>
+class SimpleSpinorBasis {
+public:
+    using ExpansionScalar = _ExpansionScalar;
+
+
+protected:
+    TransformationMatrix<ExpansionScalar> C;  // the matrix that holds the the expansion coefficients, i.e. that expresses the spinors in terms of the underlying scalar basis/bases
+
+
+public:
+
+    /*
+     *  CONSTRUCTORS
+     */
+
+    /**
+     *  @param C                    the matrix that holds the the expansion coefficients, i.e. that expresses the spinors in terms of the underlying scalar basis
+     */
+    SimpleSpinorBasis(const TransformationMatrix<ExpansionScalar>& C) :
+        C (C)
+    {}
+
+
+
+    /* 
+     *  PUBLIC METHODS
+     */
+
+    /**
+     *  @return the transformation matrix between the scalar basis and the current orbitals
+     */
+    const TransformationMatrix<ExpansionScalar>& coefficientMatrix() const { return this->C; }
+
+    /**
+     *  Rotate the spinor basis to another one using the given unitary transformation matrix
+     * 
+     *  @param U            the unitary transformation matrix that transforms both the alpha- and beta components
+     */
+    void rotate(const TransformationMatrix<ExpansionScalar>& U) {
+
+        // Check if the given matrix is actually unitary
+        if (!U.isUnitary(1.0e-12)) {
+            throw std::invalid_argument("SimpleSpinorBasis::rotate(const TransformationMatrix<ExpansionScalar>&): The given transformation matrix is not unitary.");
+        }
+
+        this->transform(U);
+    }
+
+
+    /**
+     *  Rotate the spinor basis to another one using the unitary transformation matrix that corresponds to the given Jacobi rotation parameters
+     * 
+     *  @param jacobi_rotation_parameters       the Jacobi rotation parameters (p, q, angle) that are used to specify a Jacobi rotation: we use the (cos, sin, -sin, cos) definition for the Jacobi rotation matrix
+     * 
+     *  @note this function is only available for real spinor bases because Jacobi rotation parameters generate real rotations
+     */
+    template<typename Z = ExpansionScalar>
+    enable_if_t<std::is_same<Z, double>::value> rotate(const JacobiRotationParameters& jacobi_rotation_parameters) {
+
+        const auto dim = this->simpleDimension();
+        const auto J = TransformationMatrix<double>::FromJacobi(jacobi_rotation_parameters, dim);
+        this->rotate(J);
+    }
+
+
+    /**
+     *  @return the dimension of this simple spinor basis, i.e. the dimension of the underlying coefficient matrix
+     */
+    size_t simpleDimension() const { return this->C.cols(); }
+
+    /**
+     *  Transform the spinor basis another one using the given transformation matrix
+     * 
+     *  @param T            the transformation matrix that transforms both the alpha- and beta components
+     */
+    void transform(const TransformationMatrix<ExpansionScalar>& T) {
+
+        this->C.transform(T);
+    }
+};
+
+
+}  // namespace GQCP
